pyproject.toml
[tool.poetry]
name = "cmp"
version = "0.1.0"
description = ""
authors = ["Willseed <willseed.tw@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
yt-dlp = "^2024.12.23"
ffmpeg-python = "^0.2.0"
openai = "^1.59.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
